+++ staging.sh
@@ -8,13 +8,12 @@ prefix=""
 if [ -f "/usr/local/bin/berglas" ]; then
   prefix="berglas exec --"
 fi
-suffix=""
-if [[ "$STATSD_HOST" ]]; then
-  suffix="--statsd-host ${STATSD_HOST}:${STATSD_PORT}"
+
+if [ "$GUNICORN_WORKERS" -gt 1 ];
+then
+    export PROMETHEUS_MULTIPROC_DIR="${PROMETHEUS_MULTIPROC_DIR:-$HOME/.prometheus}"
+    rm -r ${PROMETHEUS_MULTIPROC_DIR?}/* 2> /dev/null
+    mkdir -p "$PROMETHEUS_MULTIPROC_DIR"
 fi
 
-export PROMETHEUS_MULTIPROC_DIR="${PROMETHEUS_MULTIPROC_DIR:-$HOME/.prometheus}"
-rm -r ${PROMETHEUS_MULTIPROC_DIR?}/* 2> /dev/null
-mkdir -p "$PROMETHEUS_MULTIPROC_DIR"
-
-$prefix gunicorn codecov.wsgi:application --reload --workers=2 --bind 0.0.0.0:8000 --access-logfile '-' --timeout "${GUNICORN_TIMEOUT:-600}" $suffix
+$prefix gunicorn codecov.wsgi:application --reload --workers=${GUNICORN_WORKERS:-2} --threads=${GUNICORN_THREADS:-1} --worker-connections=${GUNICORN_WORKER_CONNECTIONS:-1000} --bind 0.0.0.0:8000 --access-logfile '-' --timeout "${GUNICORN_TIMEOUT:-600}" --disable-redirect-access-to-syslog --config=gunicorn.conf.py
